Filtri, prenotazioni, completamento richieste/offerte con valutazioni obbligatorie

+++ src/main/java/org/fabiojava/timebank/domain/model/Valutazione.java
@@ -16,7 +16,7 @@ public class Valutazione {
     private String commento;
     private Timestamp dataValutazione;
     private TipoValutatore tipoValutatore;
-    private Integer idPrenotazione;
+    private Long idPrenotazione;
 
     public enum TipoValutatore {
         RICHIEDENTE,
